requirements.txt
torch
torchaudio
elevenlabs
speechbrain
pydub
faster-whisper